requirements.txt
streamlit
replicate
python-dotenv
requests
librosa
numpy
matplotlib
